+++ src/main/java/cart/auth/Extractor.java
@@ -2,7 +2,7 @@ package cart.auth;
 
 import org.springframework.http.HttpHeaders;
 
-import cart.dto.UserRequest;
+import cart.dto.user.UserRequest;
 
 public interface Extractor {
     UserRequest extractUser(HttpHeaders header);
