+++ src/main/java/cart/controller/MainController.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 
-import cart.dto.ProductResponse;
+import cart.dto.product.ProductResponse;
 import cart.service.ProductService;
 
 @Controller
