+++ src/main/java/cart/dao/user/UserDao.java
@@ -1,4 +1,4 @@
-package cart.dao;
+package cart.dao.user;
 
 import java.util.List;
