+++ src/main/java/cart/dto/product/ProductMapper.java
@@ -1,4 +1,4 @@
-package cart.dto;
+package cart.dto.product;
 
 import cart.dao.entity.Product;
