+++ src/main/java/cart/dto/product/ProductResponse.java
@@ -1,4 +1,4 @@
-package cart.dto;
+package cart.dto.product;
 
 public class ProductResponse {
     private final Long id;
