+++ src/main/java/cart/dto/user/UserMapper.java
@@ -1,4 +1,4 @@
-package cart.dto;
+package cart.dto.user;
 
 import cart.dao.entity.User;
