+++ src/main/resources/templates/cart.html
@@ -35,7 +35,7 @@
         }
 
         axios.request({
-            url: '',
+            url: '/cart/products',
             headers: {
                 'Authorization': `Basic ${credentials}`
             }
