+++ src/test/java/cart/dao/JdbcProductDaoTest.java
@@ -19,6 +19,8 @@ import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.test.context.jdbc.Sql;
 
 import cart.dao.entity.Product;
+import cart.dao.product.JdbcProductDao;
+import cart.dao.product.ProductDao;
 
 @JdbcTest
 @Sql({"/dropTable.sql", "/data.sql"})
